Improvements to how the AutoUpdater waits until an internet connection is available
